nail pic set up

+++ Source/Game/BtdClass/NailMachine.cpp
@@ -12,13 +12,12 @@ namespace Btd
         if (!BallonFactory::BallonVector.empty() && GetShootTimeCounter() > GetShootDeltaTime())
         {
             Ballon targetBallon = focus();
-            //todo check in attack range
             Vector2 direction = {1, 0};
             for (int i = 0; i < 8; i++)
             {
                 Vector2 target = {
-                    static_cast<float>(GetLeft()) + direction.X,
-                    static_cast<float>(GetTop()) + direction.Y
+                    GetLeft() + direction.X,
+                    GetTop()+ direction.Y
                 };
                 Shoot(target);
                 direction = Spin45(direction);
